Initialise bash-lib repo
This commit adds the structure of the libraries and a few funcions.
Most of the code in this commit is geared towards making sure this
repo stays tested and documented in future.

The following things are checked:
    * All functions are documented
    * All functions are tested
    * All scripts are linted

See Readme for a list of the functions that are included, and for more
information on testing with BATS.

(cherry picked from commit 1a441a32f326b09aa2ad4b8d83a14cf217ceff47)